src/tests/test_deep_learning_model.py
```python
import unittest
from src.models.deep_learning_model import DeepLearningModel

class TestDeepLearningModel(unittest.TestCase):

    def setUp(self):
        self.model = DeepLearningModel()

    def test_model_initialization(self):
        self.assertIsInstance(self.model, DeepLearningModel)

    def test_generate_story(self):
        story_prompt = "A haunted house"
        user_customizations = {"characters": ["John", "Sarah"], "setting": "Abandoned mansion", "plot_twist": "The house is not haunted, but they are the ghosts"}
        generated_story = self.model.generate_story(story_prompt, user_customizations)
        self.assertIsInstance(generated_story, str)

    def test_adjust_scare_intensity(self):
        story_prompt = "A haunted house"
        user_customizations = {"characters": ["John", "Sarah"], "setting": "Abandoned mansion", "plot_twist": "The house is not haunted, but they are the ghosts"}
        scare_intensity = 5
        adjusted_story = self.model.adjust_scare_intensity(story_prompt, user_customizations, scare_intensity)
        self.assertIsInstance(adjusted_story, str)

if __name__ == '__main__':
    unittest.main()
```